+++ CHANGELOG.md
@@ -1,9 +1,15 @@
 
+## 3.2.0.0
+
+*   Add support for pretty-printing Haskell character literals.
+    [#57](https://github.com/cdepillabout/pretty-simple/pull/57)
+    Thanks again [sjakobi](https://github.com/sjakobi)!
+
 ## 3.1.1.0
 
 *   Added a `pPrintString` function for pretty-printing a `String` that is the
     output of `show`.  Implemented in
-    [54](https://github.com/cdepillabout/pretty-simple/pull/54). Thanks
+    [#54](https://github.com/cdepillabout/pretty-simple/pull/54). Thanks
     [sureyeaah](https://github.com/sureyeaah)!
 *   Fix build on GHC-7.10.3.
     [#55](https://github.com/cdepillabout/pretty-simple/pull/55).  Thanks